fix: fixing web structure and components structure
