--- apps/web/lib/client.ts
@@ -1,12 +1,0 @@
-import { ApolloClient, HttpLink, InMemoryCache } from "@apollo/client";
-import { registerApolloClient } from "@apollo/experimental-nextjs-app-support/rsc";
-import { GRAPHQL_URL } from "../app/config";
-
-export const { getClient } = registerApolloClient(() => {
-  return new ApolloClient({
-    cache: new InMemoryCache(),
-    link: new HttpLink({
-      uri:GRAPHQL_URL,
-    }),
-  });
-});
