+++ apps/web/tailwind.config.ts
@@ -1,10 +1,14 @@
 // tailwind config is required for editor support
 
-import type { Config } from "tailwindcss";
-import sharedConfig from "@packages/tailwind-config";
+import type { Config } from 'tailwindcss';
+import sharedConfig from '@packages/tailwind-config';
 
-const config: Pick<Config, "content" | "presets"> = {
-  content: ["./app/**/*.tsx"],
+const config: Pick<Config, 'content' | 'presets'> = {
+  content: [
+    './app/**/*.tsx',
+    '../../packages/components/**/*.tsx',
+    '../../packages/components/**/*.jsx',
+  ],
   presets: [sharedConfig],
 };
