+++ packages/components/index.js
@@ -1,0 +1,2 @@
+export {default as Button} from './src/Button';
+export {default as Navbar} from './src/Navbar';
