+++ packages/components/src/Button/Button.jsx
@@ -1,0 +1,9 @@
+function Button() {
+  return (
+    <a className="px-4 py-2 bg-flame-800 text-flame-900 rounded-[10px] mt-[55px] inline-block">
+      Pokedex
+    </a>
+  );
+}
+
+export default Button;
