+++ packages/components/tailwind.config.ts
@@ -1,10 +1,9 @@
-import type { Config } from "tailwindcss";
-import sharedConfig from "@packages/tailwind-config";
+import type { Config } from 'tailwindcss';
+import sharedConfig from '@packages/tailwind-config';
 
-const config: Pick<Config, "prefix" | "presets" | "content"> = {
-  content: ["./src/**/*.tsx"],
+const config: Pick<Config, 'prefix' | 'presets' | 'content'> = {
+  content: ['./src/**/*.tsx', './src/**/*.jsx'],
   presets: [sharedConfig],
-  
 };
 
 export default config;
